[IMP] sale: remove duplicate codes for price_total
The price_total field (computed) on the account.move.line model in the account module is overrid
in the sale module with a new compute method doing exactly the same thing. This override is
useless and leads to a duplicate field on the database making the ORM slower and providing
unecessary work for the migration team.

- Added on account at revision:
    https://github.com/odoo/odoo/commit/35d379cceda77dd705548e07729c89fd0e56fecf#diff-f3712147f4b25231caa371bd83265dbeR1415
- Added on sale at revision:
    https://github.com/odoo/odoo/commit/7f2c55f873da758bcc3c1de3261f4c74075575c0#diff-eb374e5f26e99054eaf8b40daaebf6f9R118