[FIX] mail: let TAB cycle through @ mention propositions
Before this commit, when the partner mention propositions dropup was visible
on discuss composer, pressing on TAB result in selecting the next focusable DOM
element after the composer input, which is usually the 'Send' or 'Log' button.

This behaviour is confusing, because users expect that TAB either autofill the
input or cycle through the propositions.

We decided to align on Slack's behaviour of TAB, which is to cycle through
the mention popositions from top to bottom.