[REM] hr_expense: remove assigned subtype and keep only classic tracking
Since d2b13cf96caf0bec64d4583cd345e1a89afb5b11 assignation is now a notification sent in the
inbox or by email to the newly-assigned responsible. It is therefore
considered as not necessary to link a subtype to the assignation itself.
Indeed having it logged in the discussion history as a classic tracking
in addition to the notification is considered as sufficient.

This commit is related to task ID 60790. Closes #23244 .